+++ src/Database/migrations/2019_05_28_054043_add_user_role_group_id_to_users_table.php
@@ -14,7 +14,7 @@ class AddUserRoleGroupIdToUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function($table) {
-            $table->integer('user_role_group_id')->unsigned()->nullable()->after('team_id');
+            $table->bigInteger('user_role_group_id')->unsigned()->nullable()->after('team_id');
             $table->foreign('user_role_group_id')->references('id')->on('user_role_groups');
         });
     }
